+++ app/projects/[projectId]/datasets/page.js
@@ -8,402 +8,32 @@ import {
   Button,
   IconButton,
   Paper,
-  Table,
-  TableBody,
-  TableCell,
-  TableContainer,
-  TableHead,
-  TableRow,
   CircularProgress,
-  Alert,
-  Snackbar,
-  Chip,
   Dialog,
   DialogTitle,
   DialogContent,
   DialogActions,
-  TablePagination,
   Card,
-  Divider,
   useTheme,
   alpha,
   InputBase,
-  Tooltip,
-  Checkbox,
   LinearProgress,
   Select,
-  MenuItem,
-  TextField
+  MenuItem
 } from '@mui/material';
 import DeleteIcon from '@mui/icons-material/Delete';
-import VisibilityIcon from '@mui/icons-material/Visibility';
 import SearchIcon from '@mui/icons-material/Search';
 import FileDownloadIcon from '@mui/icons-material/FileDownload';
 import FilterListIcon from '@mui/icons-material/FilterList';
 import { useRouter } from 'next/navigation';
 import ExportDatasetDialog from '@/components/ExportDatasetDialog';
 import { useTranslation } from 'react-i18next';
+import DatasetList from './components/DatasetList';
+import useDatasetExport from './hooks/useDatasetExport';
 import { processInParallel } from '@/lib/util/async';
 import axios from 'axios';
 import { useDebounce } from '@/hooks/useDebounce';
-
-// 数据集列表组件
-const DatasetList = ({
-  datasets,
-  onViewDetails,
-  onDelete,
-  page,
-  rowsPerPage,
-  onPageChange,
-  onRowsPerPageChange,
-  total,
-  selectedIds,
-  onSelectAll,
-  onSelectItem
-}) => {
-  const theme = useTheme();
-  const { t } = useTranslation();
-  const bgColor = theme.palette.mode === 'dark' ? theme.palette.primary.dark : theme.palette.primary.light;
-  const color =
-    theme.palette.mode === 'dark'
-      ? theme.palette.getContrastText(theme.palette.primary.main)
-      : theme.palette.getContrastText(theme.palette.primary.contrastText);
-  return (
-    <Card elevation={2}>
-      <TableContainer sx={{ overflowX: 'auto' }}>
-        <Table sx={{ minWidth: 750 }}>
-          <TableHead>
-            <TableRow>
-              <TableCell
-                padding="checkbox"
-                sx={{
-                  backgroundColor: bgColor,
-                  color: color,
-                  borderBottom: `2px solid ${theme.palette.divider}`
-                }}
-              >
-                <Checkbox
-                  color="primary"
-                  indeterminate={selectedIds.length > 0 && selectedIds.length < total}
-                  checked={total > 0 && selectedIds.length === total}
-                  onChange={onSelectAll}
-                />
-              </TableCell>
-              <TableCell
-                sx={{
-                  backgroundColor: bgColor,
-                  color: color,
-                  fontWeight: 'bold',
-                  padding: '16px 8px',
-                  borderBottom: `2px solid ${theme.palette.divider}`
-                }}
-              >
-                {t('datasets.question')}
-              </TableCell>
-              <TableCell
-                sx={{
-                  backgroundColor: bgColor,
-                  color: color,
-                  fontWeight: 'bold',
-                  padding: '16px 8px',
-                  borderBottom: `2px solid ${theme.palette.divider}`
-                }}
-              >
-                {t('datasets.createdAt')}
-              </TableCell>
-              <TableCell
-                sx={{
-                  backgroundColor: bgColor,
-                  color: color,
-                  fontWeight: 'bold',
-                  padding: '16px 8px',
-                  borderBottom: `2px solid ${theme.palette.divider}`
-                }}
-              >
-                {t('datasets.model')}
-              </TableCell>
-              <TableCell
-                sx={{
-                  backgroundColor: bgColor,
-                  color: color,
-                  fontWeight: 'bold',
-                  padding: '16px 8px',
-                  borderBottom: `2px solid ${theme.palette.divider}`
-                }}
-              >
-                {t('datasets.domainTag')}
-              </TableCell>
-              <TableCell
-                sx={{
-                  backgroundColor: bgColor,
-                  color: color,
-                  fontWeight: 'bold',
-                  padding: '16px 8px',
-                  borderBottom: `2px solid ${theme.palette.divider}`
-                }}
-              >
-                {t('datasets.cot')}
-              </TableCell>
-              <TableCell
-                sx={{
-                  backgroundColor: bgColor,
-                  color: color,
-                  fontWeight: 'bold',
-                  padding: '16px 8px',
-                  borderBottom: `2px solid ${theme.palette.divider}`
-                }}
-              >
-                {t('datasets.answer')}
-              </TableCell>
-              {/* <TableCell
-                  sx={{
-                    backgroundColor: bgColor,
-                    color: color,
-                    fontWeight: 'bold',
-                    padding: '16px 8px',
-                    borderBottom: `2px solid ${theme.palette.divider}`
-                  }}>
-                {t('datasets.chunkId')}
-              </TableCell> */}
-              <TableCell
-                sx={{
-                  backgroundColor: bgColor,
-                  color: color,
-                  fontWeight: 'bold',
-                  padding: '16px 8px',
-                  borderBottom: `2px solid ${theme.palette.divider}`
-                }}
-              >
-                {t('common.actions')}
-              </TableCell>
-            </TableRow>
-          </TableHead>
-          <TableBody>
-            {datasets.map((dataset, index) => (
-              <TableRow
-                key={dataset.id}
-                sx={{
-                  '&:nth-of-type(odd)': { backgroundColor: alpha(theme.palette.primary.light, 0.05) },
-                  '&:hover': { backgroundColor: alpha(theme.palette.primary.light, 0.1) },
-                  cursor: 'pointer'
-                }}
-                onClick={() => onViewDetails(dataset.id)}
-              >
-                <TableCell
-                  padding="checkbox"
-                  sx={{
-                    borderLeft: `4px solid ${theme.palette.primary.main}`
-                  }}
-                >
-                  <Checkbox
-                    color="primary"
-                    checked={selectedIds.includes(dataset.id)}
-                    onChange={e => {
-                      e.stopPropagation();
-                      onSelectItem(dataset.id);
-                    }}
-                    onClick={e => e.stopPropagation()}
-                  />
-                </TableCell>
-                <TableCell
-                  sx={{
-                    maxWidth: 300,
-                    whiteSpace: 'normal',
-                    wordBreak: 'break-word',
-                    py: 2
-                  }}
-                >
-                  <Typography variant="body2" fontWeight="medium">
-                    {dataset.confirmed}
-                    {dataset.confirmed && (
-                      <Chip
-                        label={t('datasets.confirmed')}
-                        size="small"
-                        sx={{
-                          backgroundColor: alpha(theme.palette.success.main, 0.1),
-                          color: theme.palette.success.dark,
-                          fontWeight: 'medium',
-                          verticalAlign: 'baseline',
-                          marginRight: '2px'
-                        }}
-                      />
-                    )}
-                    {dataset.question}
-                  </Typography>
-                </TableCell>
-                <TableCell>
-                  <Typography variant="body2" color="text.secondary">
-                    {new Date(dataset.createAt).toLocaleString('zh-CN')}
-                  </Typography>
-                </TableCell>
-                <TableCell>
-                  <Chip
-                    label={dataset.model}
-                    size="small"
-                    sx={{
-                      backgroundColor: alpha(theme.palette.info.main, 0.1),
-                      color: theme.palette.info.dark,
-                      fontWeight: 'medium'
-                    }}
-                  />
-                </TableCell>
-                <TableCell>
-                  {dataset.questionLabel ? (
-                    <Chip
-                      label={dataset.questionLabel}
-                      size="small"
-                      sx={{
-                        backgroundColor: alpha(theme.palette.primary.main, 0.1),
-                        color: theme.palette.primary.dark,
-                        fontWeight: 'medium'
-                      }}
-                    />
-                  ) : (
-                    <Typography variant="body2" color="text.disabled">
-                      {t('datasets.noTag')}
-                    </Typography>
-                  )}
-                </TableCell>
-                <TableCell>
-                  <Chip
-                    label={dataset.cot ? t('common.yes') : t('common.no')}
-                    size="small"
-                    sx={{
-                      backgroundColor: dataset.cot
-                        ? alpha(theme.palette.success.main, 0.1)
-                        : alpha(theme.palette.grey[500], 0.1),
-                      color: dataset.cot ? theme.palette.success.dark : theme.palette.grey[700],
-                      fontWeight: 'medium'
-                    }}
-                  />
-                </TableCell>
-                <TableCell sx={{ maxWidth: 200 }}>
-                  <Typography
-                    variant="body2"
-                    sx={{
-                      overflow: 'hidden',
-                      textOverflow: 'ellipsis',
-                      display: '-webkit-box',
-                      WebkitLineClamp: 2,
-                      WebkitBoxOrient: 'vertical'
-                    }}
-                  >
-                    {dataset.answer}
-                  </Typography>
-                </TableCell>
-                {/* <TableCell sx={{ maxWidth: 200 }}>
-                  <Typography
-                    variant="body2"
-                    sx={{
-                      overflow: 'hidden',
-                      textOverflow: 'ellipsis',
-                      display: '-webkit-box',
-                      WebkitLineClamp: 2,
-                      WebkitBoxOrient: 'vertical'
-                    }}>
-                    {dataset.chunkId}
-                  </Typography>
-                </TableCell> */}
-                <TableCell sx={{ width: 120 }}>
-                  <Box sx={{ display: 'flex' }}>
-                    <Tooltip title={t('datasets.viewDetails')}>
-                      <IconButton
-                        size="small"
-                        onClick={e => {
-                          e.stopPropagation();
-                          onViewDetails(dataset.id);
-                        }}
-                        sx={{
-                          color: theme.palette.primary.main,
-                          '&:hover': { backgroundColor: alpha(theme.palette.primary.main, 0.1) }
-                        }}
-                      >
-                        <VisibilityIcon fontSize="small" />
-                      </IconButton>
-                    </Tooltip>
-                    <Tooltip title={t('common.delete')}>
-                      <IconButton
-                        size="small"
-                        onClick={e => {
-                          e.stopPropagation();
-                          onDelete(dataset);
-                        }}
-                        sx={{
-                          color: theme.palette.error.main,
-                          '&:hover': { backgroundColor: alpha(theme.palette.error.main, 0.1) }
-                        }}
-                      >
-                        <DeleteIcon fontSize="small" />
-                      </IconButton>
-                    </Tooltip>
-                  </Box>
-                </TableCell>
-              </TableRow>
-            ))}
-            {datasets.length === 0 && (
-              <TableRow>
-                <TableCell colSpan={7} align="center" sx={{ py: 6 }}>
-                  <Typography variant="body1" color="text.secondary">
-                    {t('datasets.noData')}
-                  </Typography>
-                </TableCell>
-              </TableRow>
-            )}
-          </TableBody>
-        </Table>
-      </TableContainer>
-      <Divider />
-      <Box
-        sx={{
-          display: 'flex',
-          alignItems: 'center',
-          justifyContent: 'space-between',
-          px: 2,
-          py: 1,
-          borderTop: `1px solid ${theme.palette.divider}`
-        }}
-      >
-        <TablePagination
-          component="div"
-          count={total}
-          page={page - 1}
-          onPageChange={onPageChange}
-          rowsPerPage={rowsPerPage}
-          onRowsPerPageChange={onRowsPerPageChange}
-          labelRowsPerPage={t('datasets.rowsPerPage')}
-          labelDisplayedRows={({ from, to, count }) => t('datasets.pagination', { from, to, count })}
-          sx={{
-            '.MuiTablePagination-selectLabel, .MuiTablePagination-displayedRows': {
-              fontWeight: 'medium'
-            },
-            border: 'none'
-          }}
-        />
-        <Box sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
-          <Typography variant="body2">{t('common.jumpTo')}:</Typography>
-          <TextField
-            size="small"
-            type="number"
-            inputProps={{
-              min: 1,
-              max: Math.ceil(total / rowsPerPage),
-              style: { padding: '4px 8px', width: '50px' }
-            }}
-            onKeyPress={e => {
-              if (e.key === 'Enter') {
-                const pageNum = parseInt(e.target.value, 10);
-                if (pageNum >= 1 && pageNum <= Math.ceil(total / rowsPerPage)) {
-                  onPageChange(null, pageNum - 1);
-                  e.target.value = '';
-                }
-              }
-            }}
-          />
-        </Box>
-      </Box>
-    </Card>
-  );
-};
+import { toast } from 'sonner';
 
 // 删除确认对话框
 const DeleteConfirmDialog = ({ open, datasets, onClose, onConfirm, batch, progress, deleting }) => {
@@ -505,17 +135,10 @@ export default function DatasetsPage({ params }) {
   const theme = useTheme();
   const [datasets, setDatasets] = useState([]);
   const [loading, setLoading] = useState(true);
-  const [snackbar, setSnackbar] = useState({
-    open: false,
-    message: '',
-    severity: 'success'
-  });
   const [deleteDialog, setDeleteDialog] = useState({
     open: false,
     datasets: null,
-    // 是否批量删除
     batch: false,
-    // 是否正在删除
     deleting: false
   });
   const [page, setPage] = useState(1);
@@ -555,11 +178,7 @@ export default function DatasetsPage({ params }) {
       );
       setDatasets(response.data);
     } catch (error) {
-      setSnackbar({
-        open: true,
-        message: error.message,
-        severity: 'error'
-      });
+      toast.error(error.message);
     } finally {
       setLoading(false);
     }
@@ -598,15 +217,6 @@ export default function DatasetsPage({ params }) {
   };
 
   const handleBatchDeleteDataset = async () => {
-    if (selectedIds.length === 0) {
-      setSnackbar({
-        open: true,
-        message: t('datasets.noSelected'),
-        severity: 'warning'
-      });
-      return;
-    }
-
     const datasetsArray = selectedIds.map(id => ({ id }));
     setDeleteDialog({
       open: true,
@@ -664,18 +274,10 @@ export default function DatasetsPage({ params }) {
         }
       );
 
-      setSnackbar({
-        open: true,
-        message: t('datasets.batchDeleteSuccess', { count: selectedIds.length }),
-        severity: 'success'
-      });
+      toast.success(t('common.deleteSuccess'));
     } catch (error) {
       console.error('批量删除失败:', error);
-      setSnackbar({
-        open: true,
-        message: error.message || t('datasets.batchDeleteFailed'),
-        severity: 'error'
-      });
+      toast.error(error.message || t('common.deleteFailed'));
     }
   };
 
@@ -687,194 +289,27 @@ export default function DatasetsPage({ params }) {
       });
       if (!response.ok) throw new Error(t('datasets.deleteFailed'));
 
-      setSnackbar({
-        open: true,
-        message: t('datasets.deleteSuccess'),
-        severity: 'success'
-      });
+      toast.success(t('datasets.deleteSuccess'));
     } catch (error) {
-      setSnackbar({
-        open: true,
-        message: error.message,
-        severity: 'error'
-      });
+      toast.error(error.message || t('datasets.deleteFailed'));
     }
   };
 
-  // 导出数据集
+  // 使用自定义 Hook 处理数据集导出逻辑
+  const { exportDatasets } = useDatasetExport(projectId);
+
+  // 处理导出数据集
   const handleExportDatasets = async exportOptions => {
-    try {
-      let apiUrl = `/api/projects/${projectId}/datasets/export`;
-      if (exportOptions.confirmedOnly) {
-        apiUrl += `?status=confirmed`;
-      }
-      const response = await axios.get(apiUrl);
-      let dataToExport = response.data;
-
-      // 根据选择的格式转换数据
-      let formattedData;
-      // 不同文件格式
-      let mimeType = 'application/json';
-
-      if (exportOptions.formatType === 'alpaca') {
-        // 根据选择的字段类型生成不同的数据格式
-        if (exportOptions.alpacaFieldType === 'instruction') {
-          // 使用 instruction 字段
-          formattedData = dataToExport.map(({ question, answer, cot }) => ({
-            instruction: question,
-            input: '',
-            output:
-              cot && exportOptions.includeCOT
-                ? `<think>${cot}</think>
-${answer}`
-                : answer,
-            system: exportOptions.systemPrompt || ''
-          }));
-        } else {
-          // 使用 input 字段
-          formattedData = dataToExport.map(({ question, answer, cot }) => ({
-            instruction: exportOptions.customInstruction || '',
-            input: question,
-            output:
-              cot && exportOptions.includeCOT
-                ? `<think>${cot}</think>
-${answer}`
-                : answer,
-            system: exportOptions.systemPrompt || ''
-          }));
-        }
-      } else if (exportOptions.formatType === 'sharegpt') {
-        formattedData = dataToExport.map(({ question, answer, cot }) => {
-          const messages = [];
-
-          // 添加系统提示词（如果有）
-          if (exportOptions.systemPrompt) {
-            messages.push({
-              role: 'system',
-              content: exportOptions.systemPrompt
-            });
-          }
-
-          // 添加用户问题
-          messages.push({
-            role: 'user',
-            content: question
-          });
-
-          // 添加助手回答
-          messages.push({
-            role: 'assistant',
-            content: cot && exportOptions.includeCOT ? `<think>${cot}</think>\n${answer}` : answer
-          });
-
-          return { messages };
-        });
-      } else if (exportOptions.formatType === 'custom') {
-        // 处理自定义格式
-        const { questionField, answerField, cotField, includeLabels, includeChunk } = exportOptions.customFields;
-        formattedData = dataToExport.map(({ question, answer, cot, questionLabel: labels, chunkId }) => {
-          const item = {
-            [questionField]: question,
-            [answerField]: answer
-          };
-
-          // 如果有思维链且用户选择包含思维链，则添加思维链字段
-          if (cot && exportOptions.includeCOT && cotField) {
-            item[cotField] = cot;
-          }
-
-          // 如果需要包含标签
-          if (includeLabels && labels && labels.length > 0) {
-            item.label = labels.split(' ')[1];
-          }
-
-          // 如果需要包含文本块
-          if (includeChunk && chunkId) {
-            item.chunk = chunkId;
-          }
-
-          return item;
-        });
-      }
-
-      // 处理不同的文件格式
-      let content;
-      let fileExtension;
-
-      if (exportOptions.fileFormat === 'jsonl') {
-        // JSONL 格式：每行一个 JSON 对象
-        content = formattedData.map(item => JSON.stringify(item)).join('\n');
-        fileExtension = 'jsonl';
-      } else if (exportOptions.fileFormat === 'csv') {
-        // CSV 格式
-        const headers = Object.keys(formattedData[0] || {});
-        const csvRows = [
-          // 添加表头
-          headers.join(','),
-          // 添加数据行
-          ...formattedData.map(item =>
-            headers
-              .map(header => {
-                // 处理包含逗号、换行符或双引号的字段
-                let field = item[header]?.toString() || '';
-                if (exportOptions.formatType === 'sharegpt') field = JSON.stringify(item[header]);
-                if (field.includes(',') || field.includes('\n') || field.includes('"')) {
-                  field = `"${field.replace(/"/g, '""')}"`;
-                }
-                return field;
-              })
-              .join(',')
-          )
-        ];
-        content = csvRows.join('\n');
-        fileExtension = 'csv';
-      } else {
-        // 默认 JSON 格式
-        content = JSON.stringify(formattedData, null, 2);
-        fileExtension = 'json';
-      }
-      // 创建 Blob 对象
-      const blob = new Blob([content], { type: mimeType || 'application/json' });
-
-      // 创建下载链接
-      const url = URL.createObjectURL(blob);
-      const a = document.createElement('a');
-      a.href = url;
-      const formatSuffix = exportOptions.formatType === 'alpaca' ? 'alpaca' : 'sharegpt';
-      a.download = `datasets-${projectId}-${formatSuffix}-${new Date().toISOString().slice(0, 10)}.${fileExtension}`;
-
-      // 触发下载
-      document.body.appendChild(a);
-      a.click();
-
-      // 清理
-      document.body.removeChild(a);
-      URL.revokeObjectURL(url);
-
+    const success = await exportDatasets(exportOptions);
+    if (success) {
       // 关闭导出对话框
       handleCloseExportDialog();
-
-      setSnackbar({
-        open: true,
-        message: '数据集导出成功',
-        severity: 'success'
-      });
-    } catch (error) {
-      setSnackbar({
-        open: true,
-        message: '导出失败: ' + error.message,
-        severity: 'error'
-      });
     }
   };
+
   // 查看详情
   const handleViewDetails = id => {
     router.push(`/projects/${projectId}/datasets/${id}`);
-  };
-
-  // 关闭提示框
-  const handleCloseSnackbar = () => {
-    setSnackbar(prev => ({ ...prev, open: false }));
   };
 
   // 处理全选/取消全选
@@ -1072,7 +507,6 @@ ${answer}`
         progress={deleteProgress}
         deleting={deleteDialog.deleting}
       />
-      <ExportDatasetDialog open={exportDialog.open} onClose={handleCloseExportDialog} onExport={handleExportDatasets} />
 
       {/* 更多筛选对话框 */}
       <Dialog open={filterDialogOpen} onClose={() => setFilterDialogOpen(false)} maxWidth="xs" fullWidth>
@@ -1134,17 +568,6 @@ ${answer}`
           </Button>
         </DialogActions>
       </Dialog>
-
-      <Snackbar
-        open={snackbar.open}
-        autoHideDuration={2000}
-        onClose={handleCloseSnackbar}
-        anchorOrigin={{ vertical: 'bottom', horizontal: 'center' }}
-      >
-        <Alert onClose={handleCloseSnackbar} severity={snackbar.severity} variant="filled" sx={{ width: '100%' }}>
-          {snackbar.message}
-        </Alert>
-      </Snackbar>
 
       <ExportDatasetDialog
         open={exportDialog.open}
